Move pgvector prisma to common
