+++ packages/common/.npmignore
@@ -1,0 +1,1 @@
+prisma_pgvector/generated
