+++ packages/common/README.md
@@ -5,3 +5,7 @@ Modifying this:
 1. `npm login` according to credentials in TROFOS playbook
 
 2. `npm publish --access public`
+
+The main application backend's prisma client is automatically generated when this package is installed in another package. It can then be accessed from the installing package's node modules as per normal
+
+Since there is another prisma schema for the vector db, the prisma client for pgvector is has to be exported from src here, and imported accordingly.
