+++ packages/common/prisma/currentTime.ts
@@ -1,0 +1,2 @@
+export const CURRENT_YEAR = Number(process.env.CURRENT_YEAR ?? '2022');
+export const CURRENT_SEM = Number(process.env.CURRENT_SEM ?? '1');
