+++ packages/common/src/prisma/index.ts
@@ -1,0 +1,1 @@
+export * from '../../prisma_pgvector/generated/pgvector_client';
